Changes to Message class and handling
Enums updated to byte type, internal workings  changed for easier retrieval of Message Type and sender.

+++ Proftaak_Healthcare_B3/HealthcareClient/ServerConnection/IHeartrateDataReceiver.cs
@@ -2,6 +2,6 @@
 {
     public interface IHeartrateDataReceiver
     {
-        void ReceiveHeartrateData(int heartrate, HeartrateMonitor heartrateMonitor);
+        void ReceiveHeartrateData(byte heartrate, HeartrateMonitor heartrateMonitor);
     }
 }
